prplmesh: add a dependency on hostapd and PACKAGE_hostapd-openssl
Hostapd is currently needed both at build-time and run-time.

'+@PACKAGE_hostapd-openssl' is added to the dependencies because
prplMesh requires that specific variant.
Due to how OpenWrt handles this kind of dependencies, it will be
selected when selecting prplmesh, but it won't be checked when
installed as an ipkg. For this reason, '+hostapd' is also added to
the dependencies.

While we're at it, re-order the dependencies alphabetically.

Also fix the usage of spaces instead of tabs for CATEGORY.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>