hack: prplmesh: only depend on iwlwav-driver-uci if not using SDK
When we're using the SDK to compile only prplMesh, we cannot (and
don't need to) compile iwlwav-driver-uci. It doesn't compile because
it relies on some kernel configuration, and we don't build the kernel
in that case.

Use "+!IN_SDK" for the iwlwav-driver-uci dependency so that it's not
included as a dependency if we're using the SDK.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>