feeds:prplmesh:config - add hostap_iface to config file
To provide better flexibility and per platform configuration, the
slaves radio configuration is moved to the prplmesh configuration
file - which is platform-specific.

The prplmesh configuration file is shared between the following
platforms:
RAX40, Turris Omnia, GL.Inet
As such, the new configuration is used as a default configuration. If a
platform requires a different configuration - a separate feed should be
created for that platform.
E.g.:
RAX40 use interfaces wlan0 and wlan2
Turris Omnia and GL.Inet use interfaces wlan0 and wlan1

This change is a prerequisite for the prplmesh task of moving slave
radio configuration from beerocks_agent.conf to the prplmesh config
in UCI:
"https://github.com/prplfoundation/prplMesh/issues/787"

Signed-off-by: Adam Dov <adamx.dov@intel.com>