iwlwav:prplmesh:files add prplmesh update config script
The default configuration of prplmesh in openWrt is aligned to the
Turris Omnia and the GL.Inet platforms that make use of wlan0 and wlan1
as the 2 radio interfaces.
On the Rax40, on the other hand, the wlan0 and wlan1 are used
respectively.

The openWrt provides a feature for boot-time-default-configuration
scripts that are automatically triggered as part of the boot sequence.
The above-mentioned scripts must be installed to /etc/uci-defaults/ for
the script to be triggered at boot.
The sequence of the above-mentioned scripts is determined by the script
file name: X-<rest-of-file-name>, where X=2 is triggered after files
with X=1 and before files with X=3.

The added script determines which platform is booted according to the
/tmp/sysinfo/board_name. If GRX appears in the name, the platform uses
wlan2 as the second radio, and configuration is updated accordingly.

Signed-off-by: Adam Dov <adamx.dov@intel.com>